feat: add LottoNumbers and modify LottoTicket and LottoTickets
