+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,42 +1,18 @@
 package lotto.domain;
 
-import java.util.List;
 import java.util.Objects;
 
 public class LottoTicket {
-    public static final int AMOUNT_OF_NUMBERS = 6;
-    public static final int MINIMUM_NUMBER = 1;
-    public static final int MAXIMUM_NUMBER = 45;
-    private final List<Integer> numbers;
+    private final LottoNumbers lottoNumbers;
 
-    public LottoTicket(final List<Integer> numbers) {
-        validateNumbers(numbers);
-        this.numbers = numbers;
-    }
-
-    private static void validateNumbers(final List<Integer> numbers) {
-        checkAmountOfNumbers(numbers);
-        for (final int number : numbers) {
-            checkRangeOfNumber(number);
-        }
-    }
-
-    private static void checkAmountOfNumbers(final List<Integer> numbers) {
-        if (numbers.size() != AMOUNT_OF_NUMBERS) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private static void checkRangeOfNumber(final int number) {
-        if (number < MINIMUM_NUMBER || number > MAXIMUM_NUMBER) {
-            throw new IllegalArgumentException();
-        }
+    public LottoTicket(LottoNumbers lottoNumbers) {
+        this.lottoNumbers = lottoNumbers;
     }
 
     @Override
     public String toString() {
         return "LottoTicket{" +
-                "numbers=" + numbers +
+                "lottoNumbers=" + lottoNumbers +
                 '}';
     }
 
@@ -49,11 +25,11 @@ public class LottoTicket {
             return false;
         }
         final LottoTicket that = (LottoTicket) o;
-        return Objects.equals(numbers, that.numbers);
+        return Objects.equals(lottoNumbers, that.lottoNumbers);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(numbers);
+        return Objects.hash(lottoNumbers);
     }
 }
