+++ src/test/java/lotto/domain/LottoTicketsTest.java
@@ -7,14 +7,13 @@ import org.junit.jupiter.api.Test;
 
 class LottoTicketsTest {
     @Test
-    @DisplayName("LottoTickets 객체가 생성되어야 한다")
+    @DisplayName("티켓 개수를 파라미터로 LottoTickets 객체가 생성되어야 한다")
     void create() {
         // given
         final int amountOfTickets = 5;
 
         // when
         final LottoTickets lottoTickets = new LottoTickets(amountOfTickets);
-        System.out.println(lottoTickets);
 
         // when and then
         assertThat(lottoTickets).isInstanceOf(LottoTickets.class);
